refactor(debugging): split debugging manager into modules
